model_evaluation/requirements.txt
# Model Evaluation Requirements

# Vector database
pymilvus==2.3.4

# Embedding models
sentence-transformers==2.2.2
transformers>=4.30.0
torch>=2.0.0

# Azure Storage
azure-storage-blob==12.19.0

# Data processing
pandas==2.1.4
numpy==1.24.3

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
